+++ Demo/ApplozicSwiftDemoTests/Mocks/ALKConversationViewModelMock.swift
@@ -44,4 +44,15 @@ class ALKConversationViewModelMock: ALKConversationViewModel {
     override func markConversationRead() {
 
     }
+
+    override func currentConversationProfile(completion: @escaping (ALKConversationProfile?) -> ()) {
+        guard contactId != nil else {
+            completion(nil)
+            return
+        }
+        var conversationProfile = ALKConversationProfile()
+        conversationProfile.name = "demoDisplayName"
+        conversationProfile.status = ALKConversationProfile.Status(isOnline: false, lastSeenAt: nil)
+        completion(conversationProfile)
+    }
 }