+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 ### Enhancments
 - Updates Swift version to 4.2
 - Added Online/LastSeen status at navigation bar in conversation detail screen.
+- Show loading indicator at navigation bar till conversation details are fetched.
 - [AL-3206] Added support for new richMessage type : `ListTemplate` and deprecated `ALKGenericList`
 - [AL-3206] Added support for new richMessage type : `CardTemplate` and deprecated `ALKGenericCard`
 - [AL-3203] Removed the bubble image icon and added the corner Radius